backend/requirements.txt
beautifulsoup4==4.13.5
fastapi==0.117.1
fastmcp==2.12.3
httpx==0.28.1
langchain_community==0.3.29
langchain_core==0.3.76
langchain_mcp_adapters==0.1.10
langchain_openai==0.3.33
langgraph==0.4.8
markdownify==0.14.1
pydantic==2.11.9
python-dotenv==1.1.1
readabilipy==0.3.0
Requests==2.32.5
starlette==0.48.0
tavily_python==0.5.1
typing_extensions==4.15.0
uvicorn==0.37.0
